app/requirements.txt
langchain
gunicorn
boto3
openai
chromadb
tiktoken
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-distro
uvicorn
websockets
fastapi
